Consolidate database directory
