+++ tests/Storage/DatabaseEntriesRepositoryTest.php
@@ -3,7 +3,7 @@
 namespace Laravel\Telescope\Tests\Storage;
 
 use Laravel\Telescope\Storage\DatabaseEntriesRepository;
-use Laravel\Telescope\Tests\EntryModelFactory;
+use Laravel\Telescope\Database\Factories\EntryModelFactory;
 use Laravel\Telescope\Tests\FeatureTestCase;
 
 class DatabaseEntriesRepositoryTest extends FeatureTestCase
