+++ tests/Telescope/ExtractTagTest.php
@@ -5,7 +5,7 @@ namespace Laravel\Telescope\Tests\Telescope;
 use Illuminate\Mail\Mailable;
 use Laravel\Telescope\ExtractTags;
 use Laravel\Telescope\FormatModel;
-use Laravel\Telescope\Tests\EntryModelFactory;
+use Laravel\Telescope\Database\Factories\EntryModelFactory;
 use Laravel\Telescope\Tests\FeatureTestCase;
 
 class ExtractTagTest extends FeatureTestCase
